LayoutTests/platform/ios-wk2/fast/dom/focus-dialog-blur-input-type-change-crash-expected.txt

CONSOLE MESSAGE: RangeError: Maximum call stack size exceeded.
CONSOLE MESSAGE: RangeError: Maximum call stack size exceeded.
PASS
